+++ src/components/TiposDeUsuario/User/User.vue
@@ -1,0 +1,148 @@
+<template lang="pug">
+div.row.q-col-gutter-sm
+	q-input.col-xl-3.col-lg-4.col-md-6.col-sm-6.col-xs-12(
+		v-model="register.name"
+		filled
+		required
+		:rules="[val => !!val || msgCampoObr]"
+		label="Nome"
+		ref="nome"
+	)
+	q-input.col-xl-3.col-lg-4.col-md-6.col-sm-6.col-xs-12(
+		v-model.number="register.idade"
+		type="number"
+		filled
+		required
+		:rules="[val => !!val || msgCampoObr]"
+		label="Idade"
+		ref="idade"
+	)
+	q-input.col-xl-3.col-lg-4.col-md-6.col-sm-6.col-xs-12(
+		v-model="register.cpfCnpj"
+		filled
+		required
+		:rules="[val => !!val || msgCampoObr]"
+		label="CPF/CNPJ"
+		:mask="'###.###.###-##' || '##.###.###/####-##'"
+		ref="cpfCnpj"
+	)
+	q-input.col-xl-3.col-lg-4.col-md-6.col-sm-6.col-xs-12(
+		v-model="register.email"
+		filled
+		required
+		:rules="[val => !!val || msgCampoObr]"
+		label="E-mail"
+		ref="email"
+	)
+	q-input.col-xl-3.col-lg-4.col-md-6.col-sm-6.col-xs-12(
+		v-model="register.password"
+		filled
+		required
+		:type="isPwd ? 'password' : 'text'"
+		:rules="[val => !!val || msgCampoObr]"
+		label="Senha"
+		ref="password"
+		data-cy="password"
+	)
+		template(
+			v-slot:append
+		)
+			q-icon(
+				:name="isPwd ? 'visibility_off' : 'visibility'"
+				class="cursor-pointer"
+				@click="isPwd = !isPwd"
+			)
+	q-input.col-xl-3.col-lg-4.col-md-6.col-sm-6.col-xs-12(
+		filled
+		v-model="register.dtNascimento"
+		label="Data Nascimento"
+		:rules="[val => !!val || msgCampoObr]"
+		readonly
+		ref="dtNascimento"
+	)
+		template(
+			v-slot:append
+		)
+			q-icon.cursor-pointer(
+				name="event"
+			)
+				q-popup-proxy.q-mr-xl(
+					ref="qDateProxy"
+					transition-show="scale"
+					transition-hide="scale"
+				)
+					q-date(
+						v-model="register.dtNascimento"
+						:locale="myLocale"
+						today-btn
+						mask="DD/MM/YYYY"
+						navigation-min-year-month="1950/01"
+						navigation-max-year-month="2021/12"
+						@blur="() => $refs.qDateProxy.hide()"
+					)
+	q-select.col-xl-3.col-lg-4.col-md-6.col-sm-6.col-xs-12(
+		v-model="register.tipoUsuario"
+		:options="optionsTypeUser"
+		:rules="[val => !!val || msgCampoObr]"
+		ref="tipoUsuario"
+		filled
+		emit-value
+		map-options
+		required
+		label="Tipo de Usuário"
+	)
+	q-toggle.col-xl-12.col-lg-12.col-md-12.col-sm-12.col-xs-12(
+		v-model="register.authGoogle"
+		color="secondary"
+		label="Autenticado com Google?"
+		keep-color
+		checked-icon="check"
+		unchecked-icon="clear"
+	)
+	AddUser(
+		:formData="register"
+	)
+</template>
+
+<script>
+import AddUser from "./AddUser.vue"
+export default {
+	name: 'User',
+	components: {
+		AddUser
+	},
+	data () {
+		return {
+			register: {
+				name: '',
+				idade: '',
+				cpfCnpj: '',
+				email: '',
+				password: '',
+				dtNascimento: '',
+				tipoUsuario: '',
+				authGoogle: false
+			},
+			isPwd: true,
+			msgCampoObr: "Faltou aqui ó",
+			optionsTypeUser: [
+				{
+					value: 'artista',
+					label: 'Artista'
+				},
+				{
+					value: 'promotor',
+					label: 'Promotor'
+				}
+			],
+			myLocale: {
+        days: 'Domingo_Segunda_Terça_Quarta_Quinta_Sexta_Sábado'.split('_'),
+        daysShort: 'Dom_Seg_Ter_Qua_Qui_Sex_Sáb'.split('_'),
+        months: 'Janeiro_Fevereiro_Março_Abril_Maio_Junho_Julho_Agosto_Setembro_Outubro_Novembro_Dezembro'.split('_'),
+        monthsShort: 'Jan_Fev_Mar_Abr_Mai_Jun_Jul_Ago_Set_Out_Nov_Dez'.split('_'),
+        firstDayOfWeek: 1
+      }
+		}
+	}
+}
+</script>
